requirements.txt
langchain>=0.0.190
transformers>=4.29.0
faiss-cpu>=1.7.3
huggingface-hub>=0.19.3
sentence-transformers>=2.2.2
pypdf>=3.9.1
docx2txt>=0.8
pandas>=2.0.3
numpy>=1.26.0
rank-bm25>=0.2.2

# Оновлюємо gradio
gradio>=4.44.1
